1.1.1 version update (#148)
* 수집기 로그삭제 주기를 14일->7일로 변경 (#139)

* ami 장치상세팝업 안나오는 오류 수정 (#140)

* 수집기 및 서비스포털 pmgrow쪽 url 변경 및 pv 관련 api url 변경(ivtes->ivts) (#141)

pmgrow쪽 url을 최신 스펙 url로 변경
ems관련 startDt파라미터 세팅 시 timestamp를 +1 -> +1000으로 변경(0.001초->1초)

* pmgrow쪽 url 변경 및 pv 관련 api url 변경(ivtes->ivts) 오류 수정 (#143)

* pv발전량 vo 형변황(Long -> Float) (#145)

* ami장치정보에 activePower 및 무효전력 추가 (#146)

* ami장치정보에 activePower 및 무효전력 추가

* 서비스포털 ami테이블 변수명 변경

* 메뉴 br태그 추가 (#147)

* 1.1.1 version update

+++ ewp-api-agent/src/main/java/kr/co/ewp/api/controller/EnergyController.java
@@ -1059,7 +1059,7 @@ public class EnergyController {
         						  pvGen.setDeviceId(_deviceId);
         						  pvGen.setSiteId(_siteId);
         						  pvGen.setStdDate(item.getTimestamp());
-        						  pvGen.setGenVal(item.getGenEnergy().floatValue());
+        						  pvGen.setGenVal(item.getGenEnergy());
         						  pvGen.setTemp(item.getTemperature());
         						  
         						  pvGentList.add(pvGen);
@@ -1089,7 +1089,7 @@ public class EnergyController {
 	      						  pvGen.setDeviceId(_deviceId);
 	      						  pvGen.setSiteId(_siteId);
 	      						  pvGen.setStdDate(item.getTimestamp());
-	      						  pvGen.setGenVal(item.getGenEnergy().floatValue());
+	      						  pvGen.setGenVal(item.getGenEnergy());
 	      						  pvGen.setTemp(item.getTemperature());
 	      						  
 	      						  pvGentList.add(pvGen);
@@ -1105,7 +1105,7 @@ public class EnergyController {
         						  pvGen.setDeviceId(_deviceId);
         						  pvGen.setSiteId(_siteId);
         						  pvGen.setStdDate(item.getTimestamp());
-        						  pvGen.setGenVal(item.getGenEnergy().floatValue());
+        						  pvGen.setGenVal(item.getGenEnergy());
         						  pvGen.setTemp(item.getTemperature());
         						  
         						  pvGentList.add(pvGen);
